requirements.txt
dash==2.14.2
pandas==2.1.3
plotly==5.18.0
numpy==1.26.2
python-dotenv==1.0.0
